CIR-127 - Migrate the styles to the new repo (#2)
* Added bootstrap module. Added CBD/UN/UNEP images. Migrated styles.

* Removed empty files and fixed patterns

Renames `.region` to `.cus-region` in the index file.

Removes the `update.scss` file, which appears to be an empty and unused style file.

* Fix style issues and adds CBD logo SVGs

Updates the GBF target styles to use a relative path for the arrow icon.

Replaces CBD logo PNGs with SVGs for better scalability and accessibility across different languages.

---------

Co-authored-by: Rodrigo Elias <rodrigo.elias@cbd.int>