Improves i18n implementation
Migrates i18n configuration to leverage Nuxt's `useHead` for setting HTML attributes and meta tags.
This allows for dynamic updates of the `lang` and `dir` attributes based on the current locale.
Additionally, the locale files are consolidated into a single `app.json` file, simplifying the management of translations.
The social card image is added to the public directory.

Relates to CIR-217-Improve-i18n-handling